Fixed issue with the DDL javadoc element provider
Issue: The known elements after javadoc tags where not always updated
correctly.
Root Cause: The platform sometimes instantiated a new DDL Element
provider and the registered listeners where sill attached to the old one
thus the notifications got lost.
Fix: A rather ugly workaround which duplicates the interface:
Interface A is for taking the requests and forwarding them to the
singleton B
B is the original interface, but now encapsulated as a singleton